requirements.txt
python-igraph==0.7.1.post6
networkx==2.4
scipy==1.4.1
numpy==1.18.1